API: NURBS Toolbox

API: NURBS Toolbox

Module

Data Structures

Evaluation Functions/Methods

Utility and Modification Functions/Methods

Recipes for NURBS Structures

Test/Demo Functions/Methods

Private Functions/Methods

Index

Module

MagMechFEM_Matlab2Julia.NURBStoolbox — Module

NURBStoolbox

NURBS Toolbox module ported from the Matlab NURBS toolbox of Mark Spink (2000) (can currently be found on Mathworks FileExchange[1]). The Matlab package dates approximately back to the year 2000. Newer versions are written for Octave and are hosted as Octave Community package[2] (those might be compatible with Matlab).

This package is licensed under GNU General Puplic License version 2.

Exports

NURBS

NURBS1D

NURBS2D

basisfun

bspdegelev

bspderiv

bspeval

bspkntins

demo4surf

democirc

democoons

democurve

democylind

demodegelev

demodercrv

demodersrf

demoellip

demogeom

demohelix

demokntins

demoline

demorect

demorevolve

demoruled

demotorus

findspan

nrb4surf

nrbcirc

nrbcoons

nrbcylind

nrbdegelev

nrbderiv

nrbdeval

nrbeval

nrbextrude

nrbkntins

nrbline

nrbmak

nrbplot

nrbplot!

nrbrect

nrbreverse

nrbrevolve

nrbruled

nrbtestcrv

nrbtestsrf

nrbtform

nrbtransp

vecangle

veccross

vecdot

vecmag

vecmag2

vecnorm_toolbox

vecrotx

vecroty

vecrotz

vecscale

vectrans

Imports

Base

Core

DocStringExtensions

LinearAlgebra

Plots

Data Structures

MagMechFEM_Matlab2Julia.NURBStoolbox.NURBS — Type

NURBS{I<:Integer,F<:AbstractFloat}

Abstract supertype for NURBS data tayps see NURBS1D and NURBS2D.

Both curves and surfaces where represented by a structure that where compatible with the Spline Toolbox from Mathworks.

MagMechFEM_Matlab2Julia.NURBStoolbox.NURBS1D — Type

NURBS1D{I<:Integer,F<:AbstractFloat} <: NURBS{I,F}

NURBS datatype representing a NURBS curve. Constructed with nrbmak.

Note:

The control points are always converted and stored within the NURBS structure as 4D homogeneous coordinates.

Field

form::String

Type name 'B-NURBS'

dim::Integer

Dimension of the control points

number::Integer

Number of Control points

order::Integer

Order of the spline

knots::Vector{F} where F<:AbstractFloat

Knot sequence

coefs::Matrix{F} where F<:AbstractFloat

Control Points

MagMechFEM_Matlab2Julia.NURBStoolbox.NURBS2D — Type

NURBS2D{I<:Integer,F<:AbstractFloat} <: NURBS{I,F}

NURBS datatype representing a NURBS surface. Constructed with nrbmak.

Note:

The control points are always converted and stored within the NURBS structure as 4D homogeneous coordinates. For a surface the spline degree is a vector [du;dv] containing the order along the U and V directions respectively.

Field

form::String

Type name 'B-NURBS'

dim::Integer

Dimension of the control points

number::Vector{I} where I<:Integer

Number of Control points

order::Vector{I} where I<:Integer

Order of the spline

knots::Array{Vector{F}, 1} where F<:AbstractFloat

Knot sequence

coefs::Array{F, 3} where F<:AbstractFloat

Control Points

Evaluation Functions/Methods

MagMechFEM_Matlab2Julia.NURBStoolbox.basisfun — Function

basisfun(i::I,u::F,p::I,U::Vector{F}
        )::Vector{F} where {I<:Integer,F<:AbstractFloat}

Calculates the B-Spline basis function at u.

Arguments:

i: knot span from findspan

u: parametric point

p: spline degree

U: knot sequence

Output:

N: basis function vector of lenght p+1

Examples:

julia> N = basisfun(i,u,p,U)

Reference:

Algorithm A2.2 from 'The NURBS BOOK' pg70 (Les Piegl , Wayne Tiller (1997))

MagMechFEM_Matlab2Julia.NURBStoolbox.bspeval — Function

bspeval(d::I,c::Matrix{F},k::Vector{F},u::Vector{F}
       )::Matrix{F} where {I<:Integer,F<:AbstractFloat}

Evaluate a univariate B-Spline.

Arguments:

d: degree of the B-Spline

c: control points, matrix of size (dim,nc)

k: knot sequence, vector of size nk

u: parametric evaluation points, vector of size nu

Output:

p: Evaluated points, matrix of size (dim,nu)

Examples:

julia> p = bspeval(d,c,k,u)

Reference:

Algorithm A3.1 from 'The NURBS BOOK' pg82 (Les Piegl , Wayne Tiller (1997))

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbmak — Function

nrbmak(coefs::Union{Matrix{F}, Array{F,3}},
       knots::Union{Vector{F}, Vector{Vector{F}}}
      )::NURBS where {F<:AbstractFloat}

Construct the NURBS structure given the control points and the knots.

This function is used as a convenient means of constructing the NURBS data structure. Many of the other functions in the toolbox rely on the NURBS structure been correctly defined. This function not only constructs the proper structure, but also checks for consistency. The user is still free to build his own structure, in fact a few functions in the toolbox do this for convenience.

Arguments:

coefs: Control points, these can be either Cartesian or homogeneous coordinates. For a curve the control points are represented by a matrix of size (dim,nu) and for a surface by a multidimensional array of size (dim,nu,nv). Where nu is number of points along the parametric U direction, and nv the number of points along the V direction. dim is the dimension valid options are:

2 : (x,y) 2D Cartesian coordinates 3 : (x,y,z) 3D Cartesian coordinates 4 : (wx,wy,wz,w) 4D homogeneous coordinates

knots: Non-decreasing knot sequence spanning the interval [0.0;1.0]. It is assumed that the curves and surfaces are clamped to the start and end control points by knot multiplicities equal to the spline order. For a curve knots form a vector and for a surface the knots are stored by two vectors for U and V in a vector of vectors [uknots, vknots]

Output:

nurbs: NURBS data structure see NURBS1D and NURBS2D.

Note:

The control points are always converted and stored within the NURBS structure as 4D homogeneous coordinates. For a surface the spline degree is a vector [du;dv] containing the order along the U and V directions respectively.

Examples:

Construct a 2D line from [0.0;0.0] to [1.5;3.0]. For a straight line a spline of order 2 is required. Note that the knot sequence has a multiplicity of 2 at the start [0.0;0.0] and end [1.0;1.0] in order to clamp the ends.

julia> line = nrbmak([0.0 1.5; 0.0 3.0],vec([0.0 0.0 1.0 1.0]))

Construct a surface in the x-y plane i.e

^  (0.0,1.0) ------------ (1.0,1.0)
|      |                      |
| V    |                      |
|      |      Surface         |
|      |                      |
|      |                      |
|  (0.0,0.0) ------------ (1.0,0.0)
|
|------------------------------------>
                                  U

julia> coefs = cat([0.0 0.0; 0.0 1.0],[1.0 1.0; 0.0 1.0],dims=3);
julia> knots = [vec([0.0 0.0 1.0 1.]),vec([0.0 0.0 1.0 1.0])];
julia> plane = nrbmak(coefs,knots)

Reference:

related data types: NURBS1D and NURBS2D

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbeval — Function

nrbeval(nurbs::NURBS{I,F},
        tt::Union{Vector{F}, Vector{Vector{F}}, Matrix{F}} [,
        flag::Symbol=:cartesian]
       ) where {I<:Integer, F<:AbstractFloat}

Evaluate a NURBS structure at parametric points.

Evaluation of NURBS curves or surfaces at parametric points along the U and V directions. Either homogeneous coordinates are returned if the flag is set accordingly otherwise cartesian coordinates. This function utilises the function bspeval.

Arguments:

nurbs: NURBS structure

tt: parametric evaluation points either as Vector{F} or Vector{Vector{F}}

flag: flag indicating that cartesian coordinates or homogeneous coordinates should be returned (valid values: :cartesian (default) or :homogeneous)

Output:

p::Union{Matrix{F},Array{F,3}}: Evaluated points on the NURBS curve or surface as cartesian coordinates (x,y,z). If flag=:homogeneous is set the points are returned as homogeneous coordinates (wx,wy,wz).

w::Union{Vector{F},Matrix{F}}: Weights of the homogeneous coordinates of the evaluated points. Note: Only returned it the corresponding flag is set (see above).

Examples:

julia> p = nrbeval(nurbs,tt)
julia> p,w = nrbeval(nurbs,tt,:homogeneous)
julia> p = nrbeval(srf,[ut,vt])

Evaluate the NURBS circle with twenty points from 0.0 to 1.0.

julia> nrb = nrbcirc();
julia> ut = collect(range(0.0,stop=1.0,length=20))
julia> p = nrbeval(nrb,ut)

Note:

Contrary to the Matlab version does this ported version evaluates not the varargout count to switch between different return values. A symbol flag is used instead.

Reference

see also: bspeval

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbplot — Function

nrbplot(nurbs::NURBS{I,F},
        npts::Union{I, Vector{I}} [;
        plotargs... ]
       ) where {I<:Integer,F<:AbstractFloat}

Plot a NURBS curve or a surface.

Arguments:

nurbs: NURBS structure

npts: Number of evaluation points, for a surface a vector with two elements for the number of points along the U and V directions respectively.

plotargs: plot arguments to pass to the wrapped plot function

Note:

For 3D Plots nrbplot uses the Plots.pyplot() backend. Use plot arguments compatible with that backend.

Examples:

julia> nrbplot(nrb,subd)
julia> nrbplot(nrb,subd,c=:winter)

Plot the test surface with 20 points along the U direction and 30 along the V direction

julia> nrbplot(nrbtestsrf,[20;30],c=:copper)

Note:

Contrary to the Matlab version does this ported version supports no specialised input arguments just pass named plot arguments which will be passed to the wrapped plot functions.

Note:

It is also possible to overload the Julia plot() function with the mechanic Recipes.

Reference

see also: nrbeval

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbplot! — Function

nrbplot!(nurbs::NURBS{I,F},
         npts::Union{I, Vector{I}} [;
         plotargs... ]
        ) where {I<:Integer,F<:AbstractFloat}

Plot a NURBS curve or a surface.

Wrapper of nrbplot in order to introduce the Julia plot append mechanic for that plot routine.

Note:

This function serves only as wrapper for the nrbplot function for input arguments and usage look at that function.

Reference

see nrbplot

Utility and Modification Functions/Methods

MagMechFEM_Matlab2Julia.NURBStoolbox.findspan — Function

findspan(n::I,p::I,u::F,U::Vector{F})::I where {I<:Integer,F<:AbstractFloat}

Find the span of a B-Spline knot vector at a parametric point

Arguments:

n: nuber of control points -1

p: spline degree

n: parametric point

U: knot sequence

Output:

s: knot span of u in U

Examples:

julia> s = findspan(n,p,u,U)

Reference:

Algorithm A2.1 from 'The NURBS BOOK' pg68 (Les Piegl , Wayne Tiller (1997))

MagMechFEM_Matlab2Julia.NURBStoolbox.vecscale — Function

vecscale(svec::Vector{F})::Matrix{F} where {F<:AbstractFloat}

Transformation matrix for a scaling.

Returns a (4x4) Transformation matrix for scaling.

The matrix is:

[ sx   0   0   0]
[  0  sy   0   0]
[  0   0  sz   0]
[  0   0   0   1]

Arguments:

svec: A vectors defining the scaling along the x,y and z axes. i.e. [sx; sy; sy]

Output:

st: Scaling Transformation Matrix

Examples:

julia> ss = vecscale(svec)

Scale up the NURBS line [0.0;0.0;0.0] - [1.0;1.0;1.0] by 3 along the x-axis, 2 along the y-axis and 4 along the z-axis.

julia> line = nrbline(vec([0.0 0.0 0.0]),vec([1.0 1.0 1.0]));
julia> scale = vecscale(vec([3.0 2.0 4.0]));
julia> sline = nrbtform(line, scale)

Reference

see also: nrbtform

MagMechFEM_Matlab2Julia.NURBStoolbox.vectrans — Function

vectrans(tvec::Vector{F})::Matrix{F} where {F<:AbstractFloat}

Transformation matrix for a translation.

Returns a (4x4) Transformation matrix for translation.

The matrix is:

[ 1   0   0  tx ]
[ 0   1   0  ty ]
[ 0   0   1  tz ]
[ 0   0   0   1 ]

Arguments:

tvec: A vectors defining the translation along the x,y and z axes. i.e. [tx; ty; ty]

Output:

st: Translation Transformation Matrix

Examples:

julia> st = vectrans(tvec)

Translate the NURBS line [0.0;0.0;0.0] - [1.0;1.0;1.0] by 3 along the x-axis, 2 along the y-axis and 4 along the z-axis.

julia> line = nrbline(vec([0.0 0.0 0.0]),vec([1.0 1.0 1.0]));
julia> trans = vectrans(vec([3.0 2.0 4.0]));
julia> tline = nrbtform(line, trans);

Reference

see also: nrbtform

MagMechFEM_Matlab2Julia.NURBStoolbox.vecrotx — Function

vecrotx(angle::F)::Matrix{F} where {F<:AbstractFloat}

Transformation matrix for a rotation around the x axis.

Return the (4x4) Transformation matrix for a rotation about the x axis by the defined angle.

The matrix is:

[ 1         0            0          0]
[ 0     cos(angle)  -sin(angle)     0]
[ 0     sin(angle)   cos(angle)     0]
[ 0         0            0          1]

Arguments:

angle: rotation angle defined in radians

Output:

rx: (4x4) Transformation matrix.

Examples:

julia> rz = vecrotx(angle)

Rotate the NURBS line [0.0;0.0;0.0] - [3.0;3.0;3.0] by 45 degrees around the x-axis

julia> line = nrbline(vec([0.0 0.0 0.0]),vec([3.0 3.0 3.0]));
julia> rotate = vecrotx(pi/4);
julia> rline = nrbtform(line, rotate)

Reference

see also: nrbtform

MagMechFEM_Matlab2Julia.NURBStoolbox.vecroty — Function

vecroty(angle::F)::Matrix{F} where {F<:AbstractFloat}

Transformation matrix for a rotation around the y axis.

Return the (4x4) Transformation matrix for a rotation about the y axis by the defined angle.

The matrix is:

[  cos(angle)       0        sin(angle)       0]
[      0            1            0            0]
[ -sin(angle)       0        cos(angle)       0]
[      0            0            0            1]

Arguments:

angle: rotation angle defined in radians

Output:

ry: (4x4) Transformation matrix.

Examples:

julia> ry = vecroty(angle)

Rotate the NURBS line [0.0;0.0;0.0] - [3.0;3.0;3.0] by 45 degrees around the y-axis

julia> line = nrbline(vec([0.0 0.0 0.0]),vec([3.0 3.0 3.0]));
julia> rotate = vecroty(pi/4);
julia> rline = nrbtform(line, rotate)

Reference

see also: nrbtform

MagMechFEM_Matlab2Julia.NURBStoolbox.vecrotz — Function

vecrotz(angle::F)::Matrix{F} where {F<:AbstractFloat}

Transformation matrix for a rotation around the z axis.

Return the (4x4) Transformation matrix for a rotation about the z axis by the defined angle.

The matrix is:

[  cos(angle)  -sin(angle)       0          0]
[ -sin(angle)   cos(angle)       0          0]
[      0            0            1          0]
[      0            0            0          1]

Arguments:

angle: rotation angle defined in radians

Output:

rz: (4x4) Transformation matrix.

Examples:

julia> rz = vecrotz(angle)

Rotate the NURBS line [0.0;0.0;0.0] - [3.0;3.0;3.0] by 45 degrees around the z-axis

julia> line = nrbline(vec([0.0 0.0 0.0]),vec([3.0 3.0 3.0]));
julia> rotate = vecrotz(pi/4);
julia> rline = nrbtform(line, rotate)

Reference

see also: nrbtform

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbtform — Function

nrbtform(nurbs::NURBS{I,F},tmat::Matrix{F}
        )::NURBS{I,F} where {I<:Integer,F<:AbstractFloat}

Apply transformation matrix to the NURBS.

The NURBS is transform as defined a transformation matrix of size (4,4), such as a rotation, translation or change in scale. The transformation matrix can define a single transformation or multiple series of transformations. The matrix can be simple constructed by the functions vecscale, vectrans, vecrotx, vecroty and vecrotz.

Arguments:

nurbs: NURBS data structure (see nrbmak for details).

tmatrix: Transformation matrix, a matrix of size (4,4) defining a single or multiple transformations.

Output:

tnurbs: The transformed NURBS data structure.

Examples:

julia> tnurbs = nrbtform(nurbs,tmatrix)

Rotate a square by 45 degrees about the z axis.

julia> rsqr = nrbtform(nrbrect(), vecrotz(deg2rad(45)));
julia> nrbplot(rsqr,[10;10])

Reference

see also: vecscale, vectrans, vecrotx, vecroty and vecrotz

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbextrude — Function

nrbextrude(curve::NURBS1D{I,F},vector::Vector{F}
          )::NURBS2D{I,F} where {I<:Integer,F<:AbstractFloat}

Construct a NURBS surface by extruding a NURBS curve.

Constructs a NURBS surface by extruding a NURBS curve along a defined vector. The NURBS curve forms the U direction of the surface edge, and extruded along the vector in the V direction. Note NURBS surfaces cannot be extruded.

Arguments:

crv: NURBS curve to extrude, see nrbmak.

vec: Vector along which the curve is extruded.

Output:

srf: NURBS surface constructed.

Examples:

julia> srf = nrbextrude(crv,vec)

Form a hollow cylinder by extruding a circle along the z-axis.

srf = nrbextrude(nrbcirc(),[0.0;0.0;1.0]);

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbrevolve — Function

nrbrevolve(crv::NURBS1D{I,F},vec::Vector{F},
           pnt::Vector{F}=[0.0;0.0;0.0],theta::F=2.0*π
          )::NURBS2D{I,F} where {I<:Integer,F<:AbstractFloat}

Construct a NURBS surface by revolving the profile NURBS curve around an axis defined by a point and vector.

Arguments:

crv: NURBS curve to revolve, see nrbmak.

vec: Vector defining the direction of the rotation axis.

pnt: Coordinate of the point used to define the axis of rotation.

theta: Angle to revolve the curve, default 2*π.

Output:

srf: Constructed surface.

Examples:

julia> srf = nrbrevolve(crv,vec)
julia> srf = nrbrevolve(crv,vec,pnt)
julia> srf = nrbrevolve(crv,vec,pnt,ang)

Construct a sphere by rotating a semicircle around a x-axis.

julia> crv = nrbcirc(1.0,[0.0;0.0;0.0],0.0,pi);
julia> srf = nrbrevolve(crv,[1.0;0.0;0.0],[0.0;0.0;0.0]);
julia> nrbplot(srf,[20;20]);

The algorithm:

vectrans the point to the origin [0.0;0.0;0.0]

rotate the vector into alignment with the z-axis

for each control point along the curve

determine the radius and angle of control point to the z-axis

construct a circular arc in the x-y plane with this radius and start angle and sweep angle theta

combine the arc and profile, coefs and weights.

next control point

rotate and vectrans the surface back into position by reversing 1 and 2.

MagMechFEM_Matlab2Julia.NURBStoolbox.vecdot — Function

vecdot(vec1::Matrix{F},vec2::Matrix{F})::Matrix{F} where {F<:AbstractFloat}

or

vecdot(vec1::Vector{F},vec2::Vector{F})::F where {F<:AbstractFloat}

Determines scalar dot product of two vectors.

Arguments:

vec1, vec2: An array of column vectors represented by a matrix of size (dim,nv), where dim is the dimension of the vector and nv the number of vectors.

Output:

dot: Row vector of scalars, each element corresponding to the dot product of the respective components in vec1 and vec2.

Examples:

julia> dot = vecdot(vec2,vec2)

Determine the dot product of

[2.3;3.4;5.6] and [1.2;4.5;1.2]
[5.1;0.0;2.3] and [2.5;3.2;4.0]

julia> dot = vecdot([2.3 5.1; 3.4 0.0; 5.6 2.3],[1.2 2.5; 4.5 3.2; 1.2 4.0])

MagMechFEM_Matlab2Julia.NURBStoolbox.veccross — Function

veccross(vec1::Matrix{F},vec2::Matrix{F}
        )::Matrix{F} where {F<:AbstractFloat}

or

veccross(vec1::Vector{F},vec2::Vector{F}
        )::Vector{F} where {F<:AbstractFloat}

Determines cross product of two vectors.

Arguments:

vec1, vec2: An array of column vectors represented by a matrix of size (dim,nv), where dim is the dimension of the vector and nv the number of vectors.

Output:

cross: Array of column vectors, each element is corresponding to the cross product of the respective components in vec1 and vec2.

Examples:

julia> cross = veccross(vec2,vec2)

Determine the cross product of

[2.3;3.4;5.6] and [1.2;4.5;1.2]
[5.1;0.0;2.3] and [2.5;3.2;4.0]

julia> cross=veccross([2.3 5.1; 3.4 0.0; 5.6 2.3],[1.2 2.5; 4.5 3.2; 1.2 4.0])

MagMechFEM_Matlab2Julia.NURBStoolbox.vecmag — Function

vecmag(vec::Matrix{F})::Matrix{F} where {F<:AbstractFloat}

or

vecmag(vec::Vector{F})::F where {F<:AbstractFloat}

Determines the magnitude of vectors.

Arguments:

vec: An array of column vectors represented by a matrix of size (dim,nv), where dim is the dimension of the vector and nv the number of vectors.

Output:

mvec: Magnitude of the vectors, vector of size (1,nv).

Examples:

julia> mvec = vecmag(vec)

Find the magnitude of the two vectors [0.0;2.0;1.3] and [1.5;3.4;2.3].

julia> mvec = vecmag([0.0 1.5; 2.0 3.4; 1.3 2.3])

MagMechFEM_Matlab2Julia.NURBStoolbox.vecmag2 — Function

vecmag2(vec::Matrix{F})::Matrix{F} where {F<:AbstractFloat}

or

vecmag2(vec::Vector{F})::F where {F<:AbstractFloat}

Determines the squared magnitude of vectors.

Arguments:

vec: An array of column vectors represented by a matrix of size (dim,nv), where dim is the dimension of the vector and nv the number of vectors.

Output:

mvec: Squared magnitude of the vectors, vector of size (1,nv).

Examples:

julia> mvec = vecmag2(vec)

Find the squared magnitude of the two vectors [0.0;2.0;1.3] and [1.5;3.4;2.3].

julia> mvec = vecmag2([0.0 1.5; 2.0 3.4; 1.3 2.3])

MagMechFEM_Matlab2Julia.NURBStoolbox.vecnorm_toolbox — Function

vecnorm_toolbox(vec::Union{Matrix{F}, Array{F,3}}
               )::Union{Matrix{F}, Array{F,3}} where {F<:AbstractFloat}

or

vecnorm_toolbox(vec::Vector{F})::Vector{F} where {F<:AbstractFloat}

Normalises the array of vectors, returning the unit vectors.

Arguments:

vec: An array of column vectors represented by a matrix of size (dim,nv), where dim is the dimension of the vector and nv the number of vectors.

Output:

nvec: Normalised vectors, matrix the same size as vec.

Examples:

julia> nvec = vecnorm_toolbox(vec)

Normalise the two vectors [0.0;2.0;1.3] and [1.5;3.4;2.3].

julia> nvec = vecnorm_toolbox([0.0 1.5; 2.0 3.4; 1.3 2.3])

MagMechFEM_Matlab2Julia.NURBStoolbox.vecangle — Function

vecangle(num::Matrix{F},den::Matrix{F}
        )::Matrix{F} where {F<:AbstractFloat}

or

vecangle(num::F,den::F)::Vector{F} where {F<:AbstractFloat}

An alternative to atan, returning an arctangent in the range 0 to 2*π.

The components of the vector ang are the arctangent of the corresponding enties of num./den. This function is an alternative for atan, returning an angle in the range 0 to 2*π.

Arguments:

num: Numerator, vector of size (1,nv).

den: Denominator, vector of size (1,nv).

Output:

ang: Arctangents, row vector of angles.

Examples:

julia> ang = vecmag2(num,dum)

Find the atan(1.2,2.0) and atan(1.5,3.4) using vecangle.

julia> ang = vecangle([1.2 1.5], [2.0 3.4])

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbtransp — Function

nrbtransp(srf::NURBS2D{I,F}
         )::NURBS2D{I,F} where {I<:Integer,F<:AbstractFloat}

Transpose a NURBS surface, by swapping U and V directions.

Utility function that transposes a NURBS surface, by swapping U and V directions. NURBS curves cannot be transposed.

Arguments:

srf: NURBS surface, see nrbmak.

Output:

tsrf: NURBS surface with U and V diretions transposed.

Examples:

julia> tsrf = nrbtransp(srf)

MagMechFEM_Matlab2Julia.NURBStoolbox.bspdegelev — Function

bspdegelev(d::I,c::Matrix{F},k::Vector{F},t::I
          ) where {I<:Integer,F<:AbstractFloat}

Degree elevate a univariate B-Spline.

Arguments:

d: Degree of the B-Spline.

c: Control points, matrix of size (dim,nc).

k: Knot sequence, vector of size nk.

t: Raise the B-Spline degree t times.

Output:

ic: Control points of the new B-Spline.

ik: Knot vector of the new B-Spline.

Examples:

julia> ic,ik = bspdegelev(d,c,k,t)

Reference:

Algorithm A5.9 from 'The NURBS BOOK' pg206 (Les Piegl , Wayne Tiller (1997))

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbdegelev — Function

nrbdegelev(nurbs::NURBS{I,F},ntimes::Union{I, Vector{I}}
          )::NURBS{I,F} where {I<:Integer,F<:AbstractFloat}

Degree elevates the NURBS curve or surface. This function uses the B-Spline function bspdegelev.

Arguments:

nurbs: NURBS structure (curve or surface see nrbmak)

ntimes: number or pair of numbers indicating how many times the structure shall be degree elevated along U direction or V direction

Output:

enurbs: new NURBS structure with degrees elevated

Examples:

julia> ecrv = nrbdegelev(crv,utimes);
julia> esrf = nrbdegelev(srf,[utimes;vtimes]);

Increase the NURBS surface degree twice along the V direction.

julia> esrf = nrbdegelev(srf,[0;2]);

Reference

see also: bspdegelev

MagMechFEM_Matlab2Julia.NURBStoolbox.bspkntins — Function

bspkntins(d::I,c::Matrix{F},k::Vector{F},u::Vector{F}
         ) where {I<:Integer,F<:AbstractFloat}

Insert knots into a univariate B-Spline.

Arguments:

d: Degree of the B-Spline.

c: Control points, matrix of size (dim,nc).

k: Knot sequence, vector of size nk.

u: Row vector of knots to be inserted, size nu.

Output:

ic: Control points of the new B-Spline, of size (dim,nc+nu).

ik: Knot vector of the new B-Spline, of size (nk+nu).

Examples:

julia> ic,ik = bspkntins(d,c,k,u)

Reference:

Algorithm A5.4 from 'The NURBS BOOK' pg164 (Les Piegl , Wayne Tiller (1997))

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbkntins — Function

nrbkntins(nurbs::NURBS{I,F}, iknots::Union{Vector{F}, Vector{Vector{F}}}
         )::NURBS{I,F} where {I<:Integer,F<:AbstractFloat}

Insert a single or multiple knots into a NURBS curve or surface.

Inserts knots into the NURBS data structure, these can be knots at new positions or at the location of existing knots to increase the multiplicity. Note that the knot multiplicity cannot be increased beyond the order of the spline. Knots along the V direction can only inserted into NURBS surfaces, not curves that are always defined along the U direction. This function uses the B-Spline function bspkntins.

Arguments:

nurbs: NURBS structure (curve or surface see nrbmak)

iknots: vektor of knots or pair of vector of knots to insert into respective U,V directions of the structure

Output:

enurbs: new NURBS structure with degrees elevated

Examples:

julia> icrv = nrbkntins(crv,iknots);
julia> isrf = nrbkntins(srf,[iuknots,ivknots]);

Note:

No knot multiplicity will be increased beyond the order of the spline.

Insert two knots into a curve, one at 0.3 and another twice at 0.4.

julia> icrv = nrbkntins(crv, vec([0.3 0.4 0.4]))

Insert into a surface two knots as (1) into the U knot sequence and one knot into the V knot sequence at 0.5.

julia> isrf = nrbkntins(srf, [vec([0.3 0.4 0.4]),[0.5]])

Reference

see also: bspkntins

MagMechFEM_Matlab2Julia.NURBStoolbox.bspderiv — Function

bspderiv(d::I,c::Matrix{F},k::Vector{F}
        ) where {I<:Integer,F<:AbstractFloat}

Evaluate the control points and knot sequence of the derivative of a univariate B-Spline.

Arguments:

d: Degree of the B-Spline.

c: Control points, matrix of size (dim,nc).

k: Knot sequence, vector of size nk.

Output:

dc: Control points of the derivative.

dk: Knot sequence of the derivative.

Examples:

julia> dc,dk = bspderiv(d,c,k)

Reference:

derived from Algorithm A3.3 from 'The NURBS BOOK' pg98 (Les Piegl , Wayne Tiller (1997))

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbderiv — Function

nrbderiv(nurbs::NURBS{I,F}) where {I<:Integer,F<:AbstractFloat}

Construct the first derivative representation of a NURBS curve or surface.

The derivatives of a B-Spline are themselves a B-Spline of lower degree, giving an efficient means of evaluating multiple derivatives. However, although the same approach can be applied to NURBS, the situation for NURBS is more complex. I have at present restricted the derivatives to just the first. I don't claim that this implentation is the best approach, but it will have to do for now. The function returns a data struture that for a NURBS curve contains the first derivatives of the B-Spline representation. Remember that a NURBS curve is represented by a univariate B-Spline using the homogeneous coordinates. The derivative data structure can be evaluated later with the function nrbdeval.

Arguments:

nurbs: NURBS data structure, see nrbmak.

Output:

dnurbs: A data structure that represents the first derivatives of a NURBS curve or surface.

Examples:

julia> dnurbs = nrbderiv(nurbs);

See the function nrbdeval for an example.

Reference

see also: bspderiv, nrbdeval

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbdeval — Function

nrbdeval(nurbs::NURBS{I,F},
         dnurbs::Union{NURBS1D{I,F}, Vector{NURBS2D{I,f}}},
         tt::Union{Vector{F}, Vector{Vector{F}}, Matrix{F}}
        ) where {I<:Integer,F<:AbstractFloat}

Evaluation of the derivative NURBS curve or surface.

Arguments:

nurbs: NURBS structure

dnurbs: NURBS derivative representation of nurbs

tt: parametric evaluation points either as Vector{F} or Vector{Vector{F}}

Output:

pnt::Union{Matrix{F},Array{F,3}}: Evaluated points.

jac::Union{Matrix{F}, Vector{Array{F,3}}}: Evaluated first derivatives (Jacobian).

Examples:

julia> pnt, jac = nrbdeval(nurbs, dnurbs, tt)
julia> pnt, jac = nrbdeval(nurbs, dnurbs, [tu,tv])

Determine the first derivatives a NURBS curve at 9 points for 0.0 to 1.0.

julia> tt = collect(range(0.0, 1.0, length=9));
julia> dcrv = nrbderiv(crv);
julia> pnts, jac = nrbdeval(crv, dcrv, tt)

Reference

see also: nrbeval, nrbderiv

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbreverse — Function

nrbreverse(nurbs::NURBS{I,F}
        )::NURBS{I,F} where {I<:Integer,F<:AbstractFloat}

Utility function to reverse the evaluation direction of a NURBS curve or surface.

Arguments:

nurbs: NURBS data structure (see nrbmak for details).

Output:

rnurbs: The transformed NURBS data structure.

Examples:

julia> rnurbs = nrbreverse(nurbs)

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbruled — Function

nrbruled(crv1::NURBS1D{I,F},crv2::NURBS1D{I,F}
        )::NURBS2D{I,F} where {I<:Integer,F<:AbstractFloat}

Constructs a ruled surface between two NURBS curves. The ruled surface is ruled along the V direction.

Arguments:

crv1: First NURBS curve, see nrbmak.

crv2: Second NURBS curve, see nrbmak.

Output:

srf: Ruled NURBS surface.

Examples:

julia> srf = nrbruled(crv1,crv2)

Construct a ruled surface between a semicircle and a straight line.

julia> cir = nrbcirc(1.0,[0.0;0.0;0],0.0,1.0*pi);
julia> line = nrbline(vec([-1 0.5 1]),vec([1 0.5 1]));
julia> srf = nrbruled(cir,line);
julia> nrbplot(srf,[20;20]);

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbcoons — Function

nrbcoons(u1::NURBS1D{I,F},u2::NURBS1D{I,F},
         v1::NURBS1D{I,F},v2::NURBS1D{I,F}
        )::NURBS2D{I,F} where {I<:Integer,F<:AbstractFloat}

Construction of a Coons patch.

Construction of a bilinearly blended Coons surface patch from four NURBS curves that define the boundary.

The orientation of the four NURBS boundary curves.

   ^ V direction
   |
   |      u2
   ------->--------
   |              |
   |              |
v1 ^   Surface    ^ v2
   |              |
   |              |
   ------->-----------> U direction
          u1

Arguments:

u1: NURBS curve defining the bottom U direction boundary of the constructed NURBS surface.

u2: NURBS curve defining the top U direction boundary of the constructed NURBS surface.

v1: NURBS curve defining the bottom V direction boundary of the constructed NURBS surface.

v1: NURBS curve defining the top V direction boundary of the constructed NURBS surface.

Output:

srf: Coons NURBS surface patch.

Examples:

julia> srf = nrbcoons(u1, u2, v1, v2)

Define four NURBS curves and construct a Coons surface patch.

julia> pnts = [0.0  3.0  4.5  6.5 8.0 10.0;
               0.0  0.0  0.0  0.0 0.0  0.0;
               2.0  2.0  7.0  4.0 7.0  9.0];
julia> crv1 = nrbmak(pnts, vec([0 0 0 1/3 0.5 2/3 1 1 1]));
julia> pnts= [ 0.0  3.0  5.0  8.0 10.0;
              10.0 10.0 10.0 10.0 10.0;
               3.0  5.0  8.0  6.0 10.0];
julia> crv2 = nrbmak(pnts, vec([0 0 0 1/3 2/3 1 1 1]));
julia> pnts= [0.0 0.0 0.0  0.0;
              0.0 3.0 8.0 10.0;
              2.0 0.0 5.0  3.0];
julia> crv3 = nrbmak(pnts, vec([0 0 0 0.5 1 1 1]));
julia> pnts= [10.0 10.0 10.0 10.0 10.0;
               0.0   3.0  5.0  8.0 10.0;
               9.0   7.0  7.0 10.0 10.0];
julia> crv4 = nrbmak(pnts, vec([0 0 0 0.25 0.75 1 1 1]));
julia> srf = nrbcoons(crv1, crv2, crv3, crv4);
julia> nrbplot(srf,[20;20]);

Recipes for NURBS Structures

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbline — Function

nrbline()::NURBS1D

or

nrbline(p1::Vector{F},p2::Vector{F})::NURBS1D where {F<:AbstractFloat}

Construct a straight line.

Constructs NURBS data structure for a straight line. If coordinates are included the function returns a unit straight line along the x-axis.

Arguments:

p1: 2D or 3D cartesian coordinate of the start point.

p2: 2D or 3D cartesian coordinate of the end point.

Output:

crv: NURBS curve for a straight line.

Examples:

julia> crv = nrbline()
julia> crv = nrbline(p1,p2)

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbrect — Function

nrbrect()::NURBS1D

or

nrbrect(size::F)::NURBS1D where {F<:AbstractFloat}

or

nrbrect(width::F,height::F)::NURBS1D where {F<:AbstractFloat}

Construct NURBS representation of a rectangle.

Construct a rectangle or square in the x-y plane with the bottom corner at (0,0,0). If there are no arguments provided the function constructs a unit square.

Arguments:

size: Size of the square (width = height).

width: Width of the rectangle (along x-axis).

height: Height of the rectangle (along y-axis).

Output:

crv: NURBS curve for a straight line.

Examples:

julia> crv = nrbrect()
julia> crv = nrbrect(size)
julia> crv = nrbrect(width, height)

MagMechFEM_Matlab2Julia.NURBStoolbox.nrb4surf — Function

nrb4surf(p11::Vector{F},p12::Vector{F},p21::Vector{F},p22::Vector{F}
        )::NURBS2D where {F<:AbstractFloat}

Constructs a bilinear surface defined by four coordinates.

The position of the corner points

^ V direction
|
----------------
|p21        p22|
|              |
|    SRF       |
|              |
|p11        p12|
-------------------> U direction

Arguments:

p11: Cartesian coordinate of the lhs bottom corner point.

p12: Cartesian coordinate of the rhs bottom corner point.

p21: Cartesian coordinate of the lhs top corner point.

p22: Cartesian coordinate of the rhs top corner point.

Output:

srf: NURBS bilinear surface, see nrbmak.

Examples:

julia> srf = nrb4surf(p11,p12,p21,p22)

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbcirc — Function

nrbcirc(radius::F=1.0,center::Vector{F}=[0.0;0.0;0.0],
        sang::F=0.0,eang::F=2.0*π)::NURBS1D where {F<:AbstractFloat}

Constructs NURBS data structure for a circular arc in the x-y plane. If no arguments are supplied a unit circle with center [0.0;0.0] is constructed.

Angles are defined as positive in the anti-clockwise direction.

Arguments:

radius: Radius of the circle, default 1.0

center: Center of the circle, default [0.0;0.0;0.0]

sang: Start angle, default 0 degrees

eang: End angle 360 degrees

Output:

crv: NURBS curve for a circular arc.

Examples:

julia> crv = nrbcirc()
julia> crv = nrbcirc(radius)
julia> crv = nrbcirc(radius,center)
julia> crv = nrbcirc(radius,center,sang,eang)

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbcylind — Function

nrbcylind(height::F=1.0,radius::F=1.0,center::Vector{F}=[0.0;0.0;0.0],
          sang::F=0.0,eang::F=2.0*π)::NURBS2D where {F<:AbstractFloat}

Construct a cylinder or cylindrical patch by extruding a circular arc (see nrbcirc).

Arguments:

height: Height of the cylinder along the axis, default 1.0.

radius: Radius of the circle, default 1.0.

center: Center of the circle, default [0.0;0.0;0.0]

sang: Start angle relative to the origin, default 0.0.

eang: End angle relative to the origin, default 2.0*π.

Output:

srf: NURBS surface of a extruded circular arc.

Examples:

julia> crv = nrbcylind()
julia> crv = nrbcylind(height,radius)
julia> crv = nrbcylind(height,radius)
julia> crv = nrbcylind(height,radius,center)
julia> crv = nrbcylind(height,radius,center,sang,eang)

References:

see also nrbcirc

Test/Demo Functions/Methods

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbtestcrv — Function

nrbtestcrv()::NURBS1D

Construct a simple test curve.

Output:

crv NURBS data structure of a curve

Examples:

julia> crv = nrbtestcrv()

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbtestsrf — Function

nrbtestsrf()::NURBS2D

Construct a simple test surface.

Output:

srf: NURBS data structure of a surface

Examples:

julia> srf = nrbtestsrf()

MagMechFEM_Matlab2Julia.NURBStoolbox.demoline — Function

demoline()

Demonstration of a 3D straight line

Examples:

julia> demoline()

MagMechFEM_Matlab2Julia.NURBStoolbox.demorect — Function

demorect()

Demonstrate of rectangluar curve.

Examples:

julia> demorect()

MagMechFEM_Matlab2Julia.NURBStoolbox.democirc — Function

democirc()

Demonstration of a circle arcs in the x-y plane.

Examples:

julia> democirc()

MagMechFEM_Matlab2Julia.NURBStoolbox.demoellip — Function

demoellip()

Demonstration of a unit circle transformed to a inclined ellipse by first scaling, then rotating and finally translating.

Examples:

julia> demoellip()

MagMechFEM_Matlab2Julia.NURBStoolbox.democurve — Function

democurve()

Shows two simple test curves.

Examples:

julia> democurve()

MagMechFEM_Matlab2Julia.NURBStoolbox.demohelix — Function

demohelix()

Demonstration of a 3D helical curve.

Examples:

julia> demohelix()

MagMechFEM_Matlab2Julia.NURBStoolbox.democylind — Function

democylind()

Demonstration of the construction of a cylinder.

Examples:

julia> democylind()

MagMechFEM_Matlab2Julia.NURBStoolbox.demotorus — Function

demotorus()

A second demonstration of surface construction.

Examples:

julia> demotorus()

MagMechFEM_Matlab2Julia.NURBStoolbox.demorevolve — Function

demorevolve()

Demonstration of surface construction by revolving a profile curve.

Examples:

julia> demorevolve()

MagMechFEM_Matlab2Julia.NURBStoolbox.demodegelev — Function

demodegelev()

Demonstration of the degree elevation algorithm.

Examples:

julia> demodegelev()

MagMechFEM_Matlab2Julia.NURBStoolbox.demokntins — Function

demokntins()

Demonstration of the knot insertion algorithm.

Examples:

julia> demokntins()

MagMechFEM_Matlab2Julia.NURBStoolbox.demodercrv — Function

demodercrv()

Demonstrates the construction of a general curve and determine of the derivative.

Examples:

julia> demodercrv()

MagMechFEM_Matlab2Julia.NURBStoolbox.demodersrf — Function

demodersrf()

Demonstrates the construction of a general surface derivatives.

Examples:

julia> demodersrf()

MagMechFEM_Matlab2Julia.NURBStoolbox.demoruled — Function

demoruled()

Demonstration of ruled surface construction.

Examples:

julia> demoruled()

MagMechFEM_Matlab2Julia.NURBStoolbox.democoons — Function

democoons()

Construction of a bilinearly blended Coons surface.

Examples:

julia> democoons()

MagMechFEM_Matlab2Julia.NURBStoolbox.demo4surf — Function

demo4surf()

Demonstration of a bilinear surface.

Examples:

julia> demo4surf()

MagMechFEM_Matlab2Julia.NURBStoolbox.demogeom — Function

demogeom()

Demonstration of how to construct a 2D geometric shape from a piece-wise set of NURBSs.

Examples:

julia> demogeom()

Private Functions/Methods

MagMechFEM_Matlab2Julia.NURBStoolbox.bincoeff — Function

bincoeff(n::I,k::I) where{I<:Integer}

Computes the binomial coefficient.

( n )      n!
(   ) = --------
( k )   k!(n-k)!

Arguments:

n: n of n choose k

k: k of n choose k

Output:

b: binomial coefficient of n choose k

Example:

julia> b = bincoeff(n,k)

Reference:

Algorithm 6.1.6 from 'Numerical Recipes in C', 2nd Edition pg215 (William H. Press , Saul A. Teukolsky , William T. Vetterling , Brian P. Flannery (1992))

MagMechFEM_Matlab2Julia.NURBStoolbox.factln — Function

factln(n::I) where{I<:Integer}

Computes ln(n!).

Example:

julia> a = factln(n)

Index

MagMechFEM_Matlab2Julia.NURBStoolbox

MagMechFEM_Matlab2Julia.NURBStoolbox.NURBS

MagMechFEM_Matlab2Julia.NURBStoolbox.NURBS1D

MagMechFEM_Matlab2Julia.NURBStoolbox.NURBS2D

MagMechFEM_Matlab2Julia.NURBStoolbox.basisfun

MagMechFEM_Matlab2Julia.NURBStoolbox.bincoeff

MagMechFEM_Matlab2Julia.NURBStoolbox.bspdegelev

MagMechFEM_Matlab2Julia.NURBStoolbox.bspderiv

MagMechFEM_Matlab2Julia.NURBStoolbox.bspeval

MagMechFEM_Matlab2Julia.NURBStoolbox.bspkntins

MagMechFEM_Matlab2Julia.NURBStoolbox.demo4surf

MagMechFEM_Matlab2Julia.NURBStoolbox.democirc

MagMechFEM_Matlab2Julia.NURBStoolbox.democoons

MagMechFEM_Matlab2Julia.NURBStoolbox.democurve

MagMechFEM_Matlab2Julia.NURBStoolbox.democylind

MagMechFEM_Matlab2Julia.NURBStoolbox.demodegelev

MagMechFEM_Matlab2Julia.NURBStoolbox.demodercrv

MagMechFEM_Matlab2Julia.NURBStoolbox.demodersrf

MagMechFEM_Matlab2Julia.NURBStoolbox.demoellip

MagMechFEM_Matlab2Julia.NURBStoolbox.demogeom

MagMechFEM_Matlab2Julia.NURBStoolbox.demohelix

MagMechFEM_Matlab2Julia.NURBStoolbox.demokntins

MagMechFEM_Matlab2Julia.NURBStoolbox.demoline

MagMechFEM_Matlab2Julia.NURBStoolbox.demorect

MagMechFEM_Matlab2Julia.NURBStoolbox.demorevolve

MagMechFEM_Matlab2Julia.NURBStoolbox.demoruled

MagMechFEM_Matlab2Julia.NURBStoolbox.demotorus

MagMechFEM_Matlab2Julia.NURBStoolbox.factln

MagMechFEM_Matlab2Julia.NURBStoolbox.findspan

MagMechFEM_Matlab2Julia.NURBStoolbox.nrb4surf

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbcirc

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbcoons

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbcylind

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbdegelev

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbderiv

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbdeval

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbeval

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbextrude

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbkntins

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbline

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbmak

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbplot

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbplot!

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbrect

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbreverse

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbrevolve

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbruled

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbtestcrv

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbtestsrf

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbtform

MagMechFEM_Matlab2Julia.NURBStoolbox.nrbtransp

MagMechFEM_Matlab2Julia.NURBStoolbox.vecangle

MagMechFEM_Matlab2Julia.NURBStoolbox.veccross

MagMechFEM_Matlab2Julia.NURBStoolbox.vecdot

MagMechFEM_Matlab2Julia.NURBStoolbox.vecmag

MagMechFEM_Matlab2Julia.NURBStoolbox.vecmag2

MagMechFEM_Matlab2Julia.NURBStoolbox.vecnorm_toolbox

MagMechFEM_Matlab2Julia.NURBStoolbox.vecrotx

MagMechFEM_Matlab2Julia.NURBStoolbox.vecroty

MagMechFEM_Matlab2Julia.NURBStoolbox.vecrotz

MagMechFEM_Matlab2Julia.NURBStoolbox.vecscale

MagMechFEM_Matlab2Julia.NURBStoolbox.vectrans

1Penguian (2010). NURBS Toolbox by D.M. Spink https://www.mathworks.com/matlabcentral/fileexchange/26390-nurbs-toolbox-by-d-m-spink, MATLAB Central File Exchange. Retrieved April 14, 2021.

2M. Spink, D. Claxton, C. de Falco, R. Vazquez (2021-03-09) Nurbs. Octave Forge Community packages. https://octave.sourceforge.io/nurbs/index.html