Documentation: Adaptive HBS

Documentation: Adaptive HBS

Module

Functions and Methods

Utility Functions and Methods

Index

Module

MagMechFEM_Matlab2Julia.AHBS — Module

AHBS

Adaptive Hierarchical B-Spline module. Contains some functions and programs implemented for sparse adaptive hierarchical NURBS support and further manipulation.

Note:

The Matlab version had import functions for NURBS from Rhino3D and some related UI processing tools. I did not port those as the UI capabilities of Julia are still not stable enough (imo) and the Rhino3D integration is not of importance for this port.

Exports

bernsteinBasis3D

bernstein_base1D

bivariate_NURBS

calculate_knot_span

degree_elev_1D

get_bezier_extraction_ij

knot_mult_var

knots_subd_bez

oslo1_global

trivariate_NURBS

univariate_NURBS

Imports

Base

Core

DocStringExtensions

LinearAlgebra

SparseArrays

Functions and Methods

MagMechFEM_Matlab2Julia.AHBS.bernstein_base1D — Function

bernstein_base1D(p::I,ξ::Vector{F},nd::I=0
                ) where {I<:Integer,F<:AbstractFloat}

Algorithm to compute the p+1 bernstein poylnomials and Derivatives. The function is based based on Les Piegl, Wayne Tiller (1997).

Arguments:

p: polynomial degree of bernstein polynomials

ξ: evaluation positions (from -1 to 1)

nd: return derivatives up to d^nb/dξ^nb (defaults to 0)

Output:

B: the p+1 bernstein polynomials evaluated at ξ, each column stores one polynomial

dB: the first derivative of the bernstein polynomials evaluated at ξ, each row stores one polynomial (returned only if nb >= 1)

ddB: the second derivative of the bernstein polynomials evaluated at ξ, each row stores one polynomial (returned only if nb>=2)

...: and so on ... (up to p derivatives can be requested (nd<=p))

Examples:

julia> B, dB, ddB, dddB = bernstein_base1D( 3, ξ, 3)
julia> B, dB = bernstein_base1D( 3, ξ, 1)
julia> B = bernstein_base1D( 5, ξ)

Reference:

Algorithm based on 'The NURBS BOOK' pg20 (Les Piegl, Wayne Tiller (1997))

MagMechFEM_Matlab2Julia.AHBS.bernsteinBasis3D — Function

bernsteinBasis3D(p::I, q::I, r::I,
                 ξ::Vector{F}, η::Vector{F}, ψ::Vector{F},
                 nd::I=0
                ) where {I<:Integer,F<:AbstractFloat}

This function calculates the trivariate bernstein base functions on the area ξ,η,ψ in [-1,+1]. It uses the operator kron to save calculation time. Therefore the univariate base components of that base will be computed first. This method of computation requires that the ξ,η and ψ values are univariate (have no duplicates) as the mesh will be formed automatically by kron(*,kron(*,*)).

Arguments:

p: polynomial degree of bernstein polynomials in ξ

q: polynomial degree of bernstein polynomials in η

r: polynomial degree of bernstein polynomials in ψ

ξ: evaluation positions (from -1 to 1) of first direction

η: evaluation positions (from -1 to 1) of second direction

ψ: evaluation positions (from -1 to 1) of third direction

nd: return derivatives up to d^nb/dX^nb (defaults to 0) (only implemented up to nd=2)

Output:

B: the p+1 bernstein polynomials evaluated at the mesh of ξ,η and ψ, each column stores one polynomial in a linearised way

dB: the first derivatives of the bernstein polynomials, each row stores one derivative and each column one polynomial (the polynomials extend into the third array dimension in a linearised way) (returned only if nb >= 1)

ddB: the second derivatives of the bernstein polynomials evaluated at ξ, each row stores one derivative and each column one polynomial (the polynomials extend into the third array dimension in a linearised way) (returned only if nb >= 2)

Examples:

julia> B, dB, ddB = bernsteinBasis3D( 3, 2 , 4, ξ, η, ψ, 2)
julia> B, dB = bernsteinBasis3D( 3, 2, 4, ξ, η, ψ, 1)
julia> B = bernsteinBasis3D( 3, 2, 1, ξ, η, ψ)

MagMechFEM_Matlab2Julia.AHBS.univariate_NURBS — Function

univariate_NURBS(coords::Vector{F},
                 p::I,
                 knots::Vector{F},
                 points::Matrix{F},
                 weights::Vector{F}
                ) where {I<:Integer,F<:AbstractFloat}

Calculates the points of a univariate NURBS geometry. The function is based on the B-Spline algorithm of de Boor Carl Boor (1972).

Arguments:

coords: vector containing the parametric evaluation points

p: polynomial degree

knots: knot tuple

points: point matrix of all unweighted control points each column belongs to one coordinate

weights: weight vector for the points

n: number of points

Output:

C: calculated point coordinates belonging to each coordinate

Examples:

julia> C = univariate_NURBS( coords, p, knots, points, weights)

Reference:

Algorithm based on 'On Calculating with B-Splines' from Carl Boor (1972)

Further references: Les Piegl, Wayne Tiller (1997) and Mark Spink (2000)

MagMechFEM_Matlab2Julia.AHBS.bivariate_NURBS — Function

bivariate_NURBS(coords::Matrix{F},
                p::Vector{I},
                knots::Vector{Vector{F}},
                points::Matrix{F},
                weights::Vector{F},
                n::Vector{I}
               ) where {I<:Integer,F<:AbstractFloat}

Calculates the points of a bivariate NURBS geometry. The function is based on the B-Spline algorithm of de Boor Carl Boor (1972).

Arguments:

coords: matrix containing the parametric evaluation pairs so each i-th row contains the evaluation point [xi[i] eta[i]]

p: vector of the directional polynomial degrees

knots: vector of directional knot tuples

points: linearised point matrix of all unweighted control points each column belongs to one coordinate

weights: weight vector for the points

n: number of points for each direction

Output:

C: calculated point coordinates belonging to each coordinate row

Examples:

julia> C = bivariate_NURBS( coords, p, knots, points, weights, n)

Reference:

Algorithm based on 'On Calculating with B-Splines' from Carl Boor (1972)

Further references: Les Piegl, Wayne Tiller (1997) and Mark Spink (2000)

MagMechFEM_Matlab2Julia.AHBS.trivariate_NURBS — Function

trivariate_NURBS(coords::Matrix{F},
                 p::Vector{I},
                 knots::Vector{Vector{F}},
                 points::Matrix{F},
                 weights::Vector{F},
                 n::Vector{I}
                ) where {I<:Integer,F<:AbstractFloat}

Calculates the points of a trivariate NURBS geometry. The function is based on the B-Spline algorithm of de Boor Carl Boor (1972).

Arguments:

coords: matrix containing the parametric evaluation pairs so each i-th row contains the evaluation point [xi[i] eta[i] psi[i]]

p: vector of the directional polynomial degrees

knots: vector of directional knot tuples

points: linearised point matrix of all unweighted control points each column belongs to one coordinate

weights: weight vector for the points

n: number of points for each direction

Output:

C: calculated point coordinates belonging to each coordinate row

Examples:

julia> C = trivariate_NURBS( coords, p, knots, points, weights, n)

Reference:

Algorithm based on 'On Calculating with B-Splines' from Carl Boor (1972)

Further references: Les Piegl, Wayne Tiller (1997) and Mark Spink (2000)

MagMechFEM_Matlab2Julia.AHBS.oslo1_global — Function

oslo1_global(p::I,knots_old::Vector{F},knots_new::Vector{F}
            ) where {I<:Integer,F<:AbstractFloat}

Compute the global subdivision matrix for the B-spline bases belonging to the knot vectors knotsold and knotsnew.

Arguments:

p: polynomial degree of both spline bases

knots_old: old/coarse knot vector

knots_new: knots to insert (values include the old knot vector)

Output:

M: sparse subdivision matrix

Examples:

julia> M = oslo1_global( p, knots_old, knots_new)

Algorithm based on:

Discrete B-splines and subdivision techniques in computer-aided geometric design and computer graphics, from Elaine Cohen, Tom Lyche, Richard Riesenfeld (1980)

A short proof of the Oslo algorithm, from Hartmut Prautzsch (1984)

Making the OSLO Algorithm More Efficient, from T. Lyche, K. Mørken (1986)

Multi-level Bézier extraction for hierarchical local refinement of Isogeometric Analysis, from Davide D'Angella, Stefan Kollmannsberger, Ernst Rank, Alessandro Reali (2018)

Utility Functions and Methods

MagMechFEM_Matlab2Julia.AHBS.knot_mult_var — Function

knot_mult_var(p::I,knots::Vector{F}) where {I<:Integer,F<:AbstractFloat}

Helper function knot_mult for the bs2bs-algorithm. This function computes and returns the unique knots and the multiplicity of all those knots between the first and last knot of multiplicity p+1 it returns additionally the count of those unique knots.

Arguments:

p: polynomial degree for the knot vector

knots: knot vector

Output:

multiplicity: vector indicating the occurrence of the knots returned in vector unique_knots the first and last entry will be 1, ignoring the p+1 multiplicity

unique_knots: vector containing the unique knot values

knot_count: integer indicating the count of the knot entries without multiplicity

Examples:

julia> multiplicity, unique_knots, knot_count = knot_mult_var( p, knots)

Reference:

Algorithm based on Giulio Casciola, Lucia Romani (2007)

MagMechFEM_Matlab2Julia.AHBS.knots_subd_bez — Function

knots_subd_bez(p::I,knots_old::Vector{F}
              ) where {I<:Integer,F<:AbstractFloat}

Helper function to subdivide one knot vector. It will return the subdivided knot vector, the corresponding bezier knots and the associated bezier start point matrix.

Arguments:

p: spline degree of spline base

knots_old: knot vector to subdivide

Output:

knots_subdiv: subdivided knot vector

knots_bezier: bezier knot vector of the subdivided knot vector

bezier_ij: bezier extraction start point matrix associated with the created knot vector

Examples:

julia> knots_subdiv, knots_bezier, bezier_ij = knots_subd_bez( p, knots_old)

Reference:

Algorithm based on Giulio Casciola, Lucia Romani (2007)

MagMechFEM_Matlab2Julia.AHBS.get_bezier_extraction_ij — Function

get_bezier_extraction_ij(p::I,knots::Vector{F}
                        ) where {I<:Integer,F<:AbstractFloat}

This function calculates the start indices of the bezier extraction patches. It is a variation of some part of the bs2bs-algorithm. And could be replaced by it.

Arguments:

p: spline degree of spline base

knots: knot vector to subdivide

Output:

bezier_ij: bezier extraction start point indices

Examples:

julia> bezier_ij = get_bezier_extraction_ij( p, knots)

Reference:

Algorithm based on Giulio Casciola, Lucia Romani (2007)

MagMechFEM_Matlab2Julia.AHBS.calculate_knot_span — Function

calculate_knot_span(p::I,knots::Vector{F},coords::Vector{F}
                   ) where {I<:Integer,F<:AbstractFloat}

Computes the knot span for a vector of evaluation points on a knot tuple.

Arguments:

p: polynomial degree for the knot vector

knots: knot vector

coords: evaluation parameter vector

Output:

knot_span: knot span of each parametric evaluation point of coords

Examples:

julia> knot_span = calculate_knot_span( p, knots, coord)

Reference:

see also: univariate_NURBS, bivariate_NURBS and trivariate_NURBS

MagMechFEM_Matlab2Julia.AHBS.degree_elev_1D — Function

degree_elev_1D(p::I,
               t::I,
               knots::Vector{F},
              ) where {I<:Integer,F<:AbstractFloat}

The function is based on on the algorithm A5.9 in Les Piegl, Wayne Tiller (1997). The function computes the the degree elevation matrix to elevate one open B-Spline t times. The function computes also the new knot vector. The function is not optimized for speed nor memory consumption.

Arguments:

p: polynomial degree

t: elevate the polynomial degree t times

knots: unchanged knot tuple

Output:

M_elev: degree elevation matrix

knots_elev: new knot vector

Examples:

julia> M_elev, knots_elev = degree_elev_1D(p, t, knots)

Reference:

Algorithm based on 'The NURBS Book' from Les Piegl, Wayne Tiller (1997)

Further references: Mark Spink (2000)

Index

MagMechFEM_Matlab2Julia.AHBS

MagMechFEM_Matlab2Julia.AHBS.bernsteinBasis3D

MagMechFEM_Matlab2Julia.AHBS.bernstein_base1D

MagMechFEM_Matlab2Julia.AHBS.bivariate_NURBS

MagMechFEM_Matlab2Julia.AHBS.calculate_knot_span

MagMechFEM_Matlab2Julia.AHBS.degree_elev_1D

MagMechFEM_Matlab2Julia.AHBS.get_bezier_extraction_ij

MagMechFEM_Matlab2Julia.AHBS.knot_mult_var

MagMechFEM_Matlab2Julia.AHBS.knots_subd_bez

MagMechFEM_Matlab2Julia.AHBS.oslo1_global

MagMechFEM_Matlab2Julia.AHBS.trivariate_NURBS

MagMechFEM_Matlab2Julia.AHBS.univariate_NURBS